# swagger_output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "School API",
    "description": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:3000/api/"
    }
  ],
...
  "paths": {
...
    "/student": {
      "post": {
        "tags": [
          "student"
        ],
        "summary": "Create a new student",
        "description": "Create a new student record. Only accessible by admins.",
        "operationId": "createStudent",
        "requestBody": {
          "description": "Details for creating a new student",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "courses": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                },
                "required": [
                  "userId",
                  "firstName",
                  "lastName"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Student created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "student": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "enrollmentNumber": {
                          "type": "string"
                        },
                        "courses": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "title": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Access denied"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/student/{enrollmentNumber}": {
      "get": {
        "tags": [
          "student"
        ],
        "summary": "Get student details",
        "description": "Retrieve details of a student by their enrollment number. Accessible by any authenticated user.",
        "operationId": "getStudent",
        "parameters": [
          {
            "name": "enrollmentNumber",
            "in": "path",
            "description": "Enrollment number of the student to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of student details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "student": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "enrollmentNumber": {
                          "type": "string"
                        },
                        "courses": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "title": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "student"
        ],
        "summary": "Update student details",
        "description": "Update details of a student by their enrollment number. Accessible by admins or the student themselves.",
        "operationId": "updateStudent",
        "parameters": [
          {
            "name": "enrollmentNumber",
            "in": "path",
            "description": "Enrollment number of the student to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Details to update the student",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "courses": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Student updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "student": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "enrollmentNumber": {
                          "type": "string"
                        },
                        "courses": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "title": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Student not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "student"
        ],
        "summary": "Delete a student",
        "description": "Delete a student record by enrollment number. Only accessible by admins.",
        "operationId": "deleteStudent",
        "parameters": [
          {
            "name": "enrollmentNumber",
            "in": "path",
            "description": "Enrollment number of the student to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Student not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/student/list": {
      "get": {
        "tags": [
          "student"
        ],
        "summary": "List all students",
        "description": "Retrieve a paginated list of all students. Accessible by any user.",
        "operationId": "listStudents",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of students to return",
            "schema": {
              "type": "integer",
              "enum": [
                5,
                10,
                30
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of students list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "totalStudents": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "currentPage": {
                      "type": "integer"
                    },
                    "students": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "userId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "firstName": {
                            "type": "string"
                          },
                          "lastName": {
                            "type": "string"
                          },
                          "enrollmentNumber": {
                            "type": "string"
                          },
                          "courses": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "format": "uuid"
                                },
                                "title": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}